Still nothing, I really don't know why checker isn't satisfied. I put back "float" as a return type of the AsyncGenerator

+++ python_async_comprehension/0-async_generator.py
@@ -8,7 +8,7 @@ from typing import AsyncGenerator
 import asyncio
 
 
-async def async_generator() -> AsyncGenerator[int, None]:
+async def async_generator() -> AsyncGenerator[float, None]:
     """
     Asynchronous generator that loops 10 times, waits for 1 second
     in each iteration, and yields a random float between 0 and 10.
